Fix Millenium-bug (sort of)
The temporal interpolation to the regular grid initially used a days-of-year counter. However when running a job that went beyond december 31st, this counter reset to 1. As a result the interp1d method provided erroneous data. By using toordinal instead, we use days since "0001-01-01" which does not reset in a new year, thereby solving the problem